pyproject.toml
[project]
name = "plum-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Danilo Catone", email = "danilocatone@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "dependency-injector>=4.47.1",
    "fastapi[standard]>=0.115.12",
    "gradio>=5.31.0",
    "jupyter>=1.1.1",
    "langchain-community>=0.3.24",
    "langchain-deepseek>=0.1.3",
    "langchain-ollama>=0.3.3",
    "langchain-qdrant>=0.2.0",
    "langgraph>=0.4.7",
    "langgraph-checkpoint-postgres>=2.0.21",
    "ollama>=0.4.9",
    "psycopg[binary,pool]>=3.2.9",
    "pypdf>=5.5.0",
    "python-dotenv>=1.1.0",
    "qdrant-client>=1.14.2",
    "sqlalchemy>=2.0.41",
]

[project.scripts]
plum-chatbot = "plum_chatbot:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
